mule-migration-tool-library/src/test/java/com/mulesoft/tools/migration/library/tools/TemplateParserTest.java
/*
 * Copyright (c) 2020 MuleSoft, Inc. This software is protected under international
 * copyright law. All use of this software is subject to MuleSoft's Master Subscription
 * Agreement (or other master license agreement) separately entered into in writing between
 * you and MuleSoft. If such an agreement is not in place, you may not use the software.
 *
 * Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
 *
 * 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
 *
 * 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
 *
 * 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
package com.mulesoft.tools.migration.library.tools;

import org.junit.Before;
import org.junit.Test;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.junit.Assert.*;

public class TemplateParserTest {

  private TemplateParser parser;
  private static final TemplateParser.ScriptTranslator IDENTITY_TRANSLATOR = (x) -> x;

  @Before
  public void setUp() {
    parser = TemplateParser.createMuleStyleParser();
  }

  @Test
  public void translateContainingSimpleQuotes() {
    String template = "Hello #[greeting] 'World'";
    String result = parser.translate(template, IDENTITY_TRANSLATOR);
    assertThat("Translated template is not the expected", result, equalTo("#[\"Hello $(greeting) 'World'\"]"));
  }

  @Test
  public void translateQuotedInterpolatedString() {
    String template = "INSERT INTO PLANET(POSITION, NAME) VALUES (777, '#[payload]')";
    String result = parser.translate(template, IDENTITY_TRANSLATOR);
    assertThat("Translated template is not the expected", result,
               equalTo("#[\"INSERT INTO PLANET(POSITION, NAME) VALUES (777, '$(payload)')\"]"));
  }
}
